+++ pypit/armlsd.py
@@ -12,10 +12,13 @@ from pypit import arflux
 from pypit import arload
 from pypit import armasters
 from pypit import armbase
+from pypit import arpixels
 from pypit import arproc
 from pypit import arsave
 from pypit import arsetup
 from pypit import artrace
+from pypit import artraceslits
+from pypit import traceslits
 from pypit import ardebug as debugger
 
 
@@ -129,13 +132,16 @@ def ARMLSD(fitsdict, reuseMaster=False, reloadMaster=True):
             if ('trace'+settings.argflag['reduce']['masters']['setup'] not in settings.argflag['reduce']['masters']['loaded']):
                 ###############
                 # Determine the edges of the spectrum (spatial)
-                #lordloc, rordloc, extord = artrace.trace_slits(slf, slf._mstrace[det-1], det, pcadesc="PCA trace of the slit edges")
-                lordloc, rordloc, extord = artrace.driver_trace_slits(slf._mstrace[det-1],
-                                                                      slf._pixlocn[det-1],
-                                                                      det=det,
-                                                                      settings=settings.argflag,
-                                                                      binbpx=slf._bpix[det-1],
-                                                                      armlsd=True)
+                #lordloc, rordloc, extord = artrace.driver_trace_slits(slf._mstrace[det-1],
+                #                                                      slf._pixlocn[det-1],
+                #                                                      det=det,
+                #                                                      settings=settings.argflag,
+                #                                                      binbpx=slf._bpix[det-1],
+                #                                                      armlsd=True)
+                tslits = traceslits.TraceSlits(slf._mstrace[det-1], slf._pixlocn[det-1],
+                                               det=det, settings=settings.argflag, binbpx=slf._bpix[det-1])
+                lordloc, rordloc, extord = tslits.run(armlsd=True)
+
                 slf.SetFrame(slf._lordloc, lordloc, det)
                 slf.SetFrame(slf._rordloc, rordloc, det)
                 # Initialize maskslit
@@ -143,10 +149,10 @@ def ARMLSD(fitsdict, reuseMaster=False, reloadMaster=True):
 
                 # Convert physical trace into a pixel trace
                 msgs.info("Converting physical trace locations to nearest pixel")
-                pixcen = artrace.phys_to_pix(0.5*(slf._lordloc[det-1]+slf._rordloc[det-1]), slf._pixlocn[det-1], 1)
+                pixcen = arpixels.phys_to_pix(0.5*(slf._lordloc[det-1]+slf._rordloc[det-1]), slf._pixlocn[det-1], 1)
                 pixwid = (slf._rordloc[det-1]-slf._lordloc[det-1]).mean(0).astype(np.int)
-                lordpix = artrace.phys_to_pix(slf._lordloc[det-1], slf._pixlocn[det-1], 1)
-                rordpix = artrace.phys_to_pix(slf._rordloc[det-1], slf._pixlocn[det-1], 1)
+                lordpix = arpixels.phys_to_pix(slf._lordloc[det-1], slf._pixlocn[det-1], 1)
+                rordpix = arpixels.phys_to_pix(slf._rordloc[det-1], slf._pixlocn[det-1], 1)
                 slf.SetFrame(slf._pixcen, pixcen, det)
                 slf.SetFrame(slf._pixwid, pixwid, det)
                 slf.SetFrame(slf._lordpix, lordpix, det)
@@ -160,7 +166,7 @@ def ARMLSD(fitsdict, reuseMaster=False, reloadMaster=True):
 #                arqa.slit_trace_qa(slf, slf._mstrace[det-1], slf._lordpix[det-1],
 #                                       slf._rordpix[det-1], extord,
 #                                       desc="Trace of the slit edges D{:02d}".format(det), use_slitid=det)
-                artrace.slit_trace_qa(slf, slf._mstrace[det-1], slf._lordpix[det-1],
+                artraceslits.slit_trace_qa(slf, slf._mstrace[det-1], slf._lordpix[det-1],
                                       slf._rordpix[det-1], extord,
                                       desc="Trace of the slit edges D{:02d}".format(det),
                                       use_slitid=det)
